Cast another size_t-derived result to unsigned long and use %lu instead of %d to printf() it, cleaning up a compile warning.
Monotone-Parent: d60692c3d82391bd01e15ca64725e263c1dc79ab
Monotone-Revision: ed83c6e4de8d6bf1ab7be5841fc11e8a16fb8677

+++ util.c
@@ -476,8 +476,9 @@ char *ExpandFilename(char *name)
     newname = (char *) malloc (HomeLen + strlen(name) + 2);
     if (!newname) {
 	fprintf (stderr, 
-		 "%s:  unable to allocate %d bytes to expand filename %s/%s\n",
-		 ProgramName, HomeLen + strlen(name) + 2, Home, &name[1]);
+		 "%s:  unable to allocate %lu bytes to expand filename %s/%s\n",
+		 ProgramName, (unsigned long) HomeLen + strlen(name) + 2,
+		 Home, &name[1]);
     } else {
 	(void) sprintf (newname, "%s/%s", Home, &name[1]);
     }
